moves all routes to respective router.js files, creates a master router at the api root level and changes all action fetch calls to hit the right routes
