+++ public/login.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <h3>Login:</h3>
-    <form action="/login" method="POST">
+    <form action="/api/auth/login" method="POST">
       <label for="email">Email:</label><br />
       <input type="text" name="email" />
