+++ server/api/habits/index.js
@@ -1,24 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-router.get("/:id", (req, res) => {
-  //deprecated habit route
-  console.log("new habit route, works with postman");
-  return req.database.Item.where({ id: req.params.id })
-    .fetch({
-      withRelated: ["user", "images", "category", "condition", "status"]
-    })
-    .then(results => {
-      if (results.toJSON().length === 0) {
-        throw new Error("Page not found!");
-      } else {
-        res.json(results);
-      }
-    });
-});
-
-router.get("/:category", (req, res) => {
-  console.log("new habit route, works with postman");
+router.get("/cat/:category", (req, res) => {
   let reqWithRelated;
 
   if (req.params.category === "all") {
@@ -44,6 +27,36 @@ router.get("/:category", (req, res) => {
         return result.category.category;
       });
     })
+    .then(results => {
+      return res.json(results);
+    })
+    .catch(err => {
+      console.log(err);
+      return res.json(err);
+    });
+});
+
+router.get("/:id", (req, res) => {
+  return req.database.Item.where({ id: req.params.id })
+    .fetch({
+      withRelated: ["user", "images", "category", "condition", "status"]
+    })
+    .then(results => {
+      if (results.toJSON().length === 0) {
+        throw new Error("Page not found!");
+      } else {
+        res.json(results);
+      }
+    })
+    .catch(err => {
+      console.log(err);
+      return res.json(err);
+    });
+});
+
+router.post("/", (req, res) => {
+  return req.database.Item.forge(req.body)
+    .save()
     .then(results => {
       return res.json(results);
     })
