+++ server/api/index.js
@@ -1,0 +1,9 @@
+const habits = require("../api/habits/index");
+const auth = require("../api/auth/index");
+const users = require("../api/users/index");
+
+module.exports = {
+  habits: habits,
+  auth: auth,
+  users: users
+};
